pyproject.toml
[project]
name = "fastmpc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dependency-injector>=4.48.1",
    "fastapi[standard]>=0.116.1",
    "fastmcp>=2.10.6",
    "google-genai>=1.26.0",
]
